+++ src/ohgo/models/__init__.py
@@ -6,3 +6,8 @@ from .incident import Incident
 from .weather_sensor_site import WeatherSensorSite
 from .construction import Construction
 from .query_params import QueryParams, DigitalSignParams, ConstructionParams, WeatherSensorSiteParams
+from .results.ohgo_results import CameraItemResult, CameraListResult, TravelDelayItemResult, TravelDelayListResult, \
+    DangerousSlowdownItemResult, DangerousSlowdownListResult, DigitalSignItemResult, DigitalSignListResult, \
+    IncidentItemResult, IncidentListResult, WeatherSensorSiteItemResult, WeatherSensorSiteListResult, \
+    ConstructionItemResult, ConstructionListResult
+from .results.http_results import Result, CachedResult
